+++ config/network.ts
@@ -25,13 +25,14 @@ export const networkConfig: Record<string, NetworkType> = {
     decimals: '4',
     gasPerDataByte: '1500',
     walletConnectDeepLink:
-      'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://maiar.com/',
+      'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://xportal.com/',
     walletConnectBridgeAddresses: ['https://bridge.walletconnect.org'],
-    walletAddress: 'https://devnet-wallet.elrond.com',
+    walletAddress: 'https://devnet-wallet.multiversx.com',
     apiAddress:
-      process.env.NEXT_PUBLIC_MULTIVERSX_API || 'https://devnet-api.elrond.com',
-    explorerAddress: 'https://devnet-explorer.elrond.com',
-    elrondIPFSGateway: 'https://devnet-media.elrond.com/nfts/asset/',
+      process.env.NEXT_PUBLIC_MULTIVERSX_API ||
+      'https://devnet-api.multiversx.com',
+    explorerAddress: 'https://devnet-explorer.multiversx.com',
+    multiversIPFSGateway: 'https://devnet-media.multiversx.com/nfts/asset/',
     apiTimeout: '10000',
   },
 
@@ -44,14 +45,14 @@ export const networkConfig: Record<string, NetworkType> = {
     decimals: '4',
     gasPerDataByte: '1500',
     walletConnectDeepLink:
-      'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://maiar.com/',
+      'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://xportal.com/',
     walletConnectBridgeAddresses: ['https://bridge.walletconnect.org'],
-    walletAddress: 'https://testnet-wallet.elrond.com',
+    walletAddress: 'https://testnet-wallet.multiversx.com',
     apiAddress:
       process.env.NEXT_PUBLIC_MULTIVERSX_API ||
-      'https://testnet-api.elrond.com',
-    explorerAddress: 'https://testnet-explorer.elrond.com',
-    elrondIPFSGateway: 'https://testnet-media.elrond.com/nfts/asset/',
+      'https://testnet-api.multiversx.com',
+    explorerAddress: 'https://testnet-explorer.multiversx.com',
+    multiversIPFSGateway: 'https://testnet-media.multiversx.com/nfts/asset/',
     apiTimeout: '10000',
   },
 
@@ -64,13 +65,13 @@ export const networkConfig: Record<string, NetworkType> = {
     decimals: '4',
     gasPerDataByte: '1500',
     walletConnectDeepLink:
-      'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://maiar.com/',
+      'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://xportal.com/',
     walletConnectBridgeAddresses: ['https://bridge.walletconnect.org'],
-    walletAddress: 'https://wallet.elrond.com',
+    walletAddress: 'https://wallet.multiversx.com',
     apiAddress:
-      process.env.NEXT_PUBLIC_MULTIVERSX_API || 'https://api.elrond.com',
-    explorerAddress: 'https://explorer.elrond.com',
-    elrondIPFSGateway: 'https://media.elrond.com/nfts/asset/',
+      process.env.NEXT_PUBLIC_MULTIVERSX_API || 'https://api.multiversx.com',
+    explorerAddress: 'https://explorer.multiversx.com',
+    multiversIPFSGateway: 'https://media.multiversx.com/nfts/asset/',
     apiTimeout: '10000',
   },
 };